feat(formula-fields): UI improvements - consolidate help text to tooltips, remove clutter, add guidance icons
