+++ client/src/pages/spools/list.tsx
@@ -36,7 +36,7 @@ import {
   useSpoolmanMaterials,
 } from "../../components/otherModels";
 import { removeUndefined } from "../../utils/filtering";
-import { ComplexFieldSurface, EntityType, useGetDerivedFields, useGetFields } from "../../utils/queryFields";
+import { FormulaFieldSurface, EntityType, useGetDerivedFields, useGetFields } from "../../utils/queryFields";
 import { TableState, useInitialTableState, useSavedState, useStoreInitialState } from "../../utils/saveload";
 import { useCurrencyFormatter } from "../../utils/settings";
 import { setSpoolArchived, useSpoolAdjustModal } from "./functions";
@@ -181,16 +181,14 @@ export const SpoolList = () => {
   );
   const liveDataSource = useLiveify("spool", queryDataSource, collapseSpool);
   const listFormulaFields = useMemo(
-    () => getFormulaFieldsForSurface(formulaFields.data, ComplexFieldSurface.list),
+    () => getFormulaFieldsForSurface(formulaFields.data, FormulaFieldSurface.list),
     [formulaFields.data],
   );
-  const toggleableListFormulaFields = useMemo(
-    () => listFormulaFields.filter((field) => field.allow_list_column_toggle),
-    [listFormulaFields],
-  );
+  // All list-surface formula fields are eligible for hide/show in the column picker,
+  // so we map every list formula to its derived column key here.
   const toggleableDerivedColumnKeys = useMemo(
-    () => toggleableListFormulaFields.map((field) => `derived.${field.key}`),
-    [toggleableListFormulaFields],
+    () => listFormulaFields.map((field) => `derived.${field.key}`),
+    [listFormulaFields],
   );
   const allColumnsWithExtraFields = useMemo(
     () => [
@@ -295,6 +293,8 @@ export const SpoolList = () => {
   };
 
   const updateColumnSelections = (selectedKeys: string[]) => {
+    // Persist core column visibility separately from derived-column visibility so
+    // derived keys can be toggled without rewriting the base showColumns state.
     setShowColumns(selectedKeys.filter((key) => !toggleableDerivedColumnKeys.includes(key)));
     setHiddenDerivedColumns(toggleableDerivedColumnKeys.filter((key) => !selectedKeys.includes(key)));
   };
@@ -344,7 +344,7 @@ export const SpoolList = () => {
                   };
                 }
                 if (column_id.indexOf("derived.") === 0) {
-                  const formulaField = toggleableListFormulaFields.find((field) => `derived.${field.key}` === column_id);
+                  const formulaField = listFormulaFields.find((field) => `derived.${field.key}` === column_id);
                   return {
                     key: column_id,
                     label: formulaField?.name ?? column_id,
@@ -510,7 +510,7 @@ export const SpoolList = () => {
           ...listFormulaFields.map(
             (field) => {
               const derivedColumnKey = `derived.${field.key}`;
-              if (field.allow_list_column_toggle && hiddenDerivedColumns.includes(derivedColumnKey)) {
+              if (hiddenDerivedColumns.includes(derivedColumnKey)) {
                 return undefined;
               }
