+++ client/src/pages/spools/show.tsx
@@ -10,7 +10,7 @@ import { NumberFieldUnit } from "../../components/numberField";
 import SpoolIcon from "../../components/spoolIcon";
 import { buildFormulaValues, formatFormulaValue, getFormulaFieldsForSurface } from "../../utils/formulaFields";
 import { enrichText } from "../../utils/parsing";
-import { ComplexFieldSurface, EntityType, useGetDerivedFields, useGetFields } from "../../utils/queryFields";
+import { FormulaFieldSurface, EntityType, useGetDerivedFields, useGetFields } from "../../utils/queryFields";
 import { useCurrencyFormatter } from "../../utils/settings";
 import { getBasePath } from "../../utils/url";
 import { IFilament } from "../filaments/model";
@@ -36,7 +36,7 @@ export const SpoolShow = () => {
 
   const record = data?.data;
   const showFormulaFields = useMemo(
-    () => getFormulaFieldsForSurface(formulaFields.data, ComplexFieldSurface.show),
+    () => getFormulaFieldsForSurface(formulaFields.data, FormulaFieldSurface.show),
     [formulaFields.data],
   );
   const derivedValues = useMemo(
@@ -234,7 +234,7 @@ export const SpoolShow = () => {
       {extraFields?.data?.map((field, index) => (
         <ExtraFieldDisplay key={index} field={field} value={record?.extra[field.key]} />
       ))}
-      {showFormulaFields.length > 0 && <Title level={4}>{t("settings.complex_fields.formula.header")}</Title>}
+      {showFormulaFields.length > 0 && <Title level={4}>{t("settings.formula_fields.formula.header")}</Title>}
       {showFormulaFields.map((field) => (
         <Fragment key={field.key}>
           <Title level={5}>{field.name}</Title>
